优化

+++ common/src/main/java/com/zb/pojo/Color.java
@@ -15,13 +15,13 @@ import java.util.Date;
 @AllArgsConstructor
 public class Color {
 
-  private Integer id;
+  private Long id;
   /** 颜色名称*/
   private String colorName;
 
-  private Integer createUserId;
+  private Long createUserId;
   private Date createTime;
-  private Integer userRenewId;
+  private Long userRenewId;
   private Date renewTime;
   private String token;
 
@@ -33,11 +33,11 @@ public class Color {
     this.token = token;
   }
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
 
@@ -49,11 +49,11 @@ public class Color {
     this.colorName = colorName;
   }
 
-  public Integer getCreateUserId() {
+  public Long getCreateUserId() {
     return createUserId;
   }
 
-  public void setCreateUserId(Integer createUserId) {
+  public void setCreateUserId(Long createUserId) {
     this.createUserId = createUserId;
   }
 
@@ -65,11 +65,11 @@ public class Color {
     this.createTime = createTime;
   }
 
-  public Integer getUserRenewId() {
+  public Long getUserRenewId() {
     return userRenewId;
   }
 
-  public void setUserRenewId(Integer userRenewId) {
+  public void setUserRenewId(Long userRenewId) {
     this.userRenewId = userRenewId;
   }
