+++ info/src/main/java/com/zb/service/PriceService.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 public interface PriceService {
     //查询
-    List<Price> listPrice(Integer sku_id);
+    List<Price> listPrice(Long sku_id);
 }
